requirements.txt
# packages in environment at /home/rick/anaconda3/envs/pytorch:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
absl-py                   0.7.1                    pypi_0    pypi
albumentations            0.1.8                    py37_0    albumentations
blas                      1.0                         mkl  
bzip2                     1.0.8                h7b6447c_0  
ca-certificates           2019.6.16            hecc5488_0    conda-forge
cairo                     1.14.12              h8948797_3  
certifi                   2019.6.16                py37_1    conda-forge
cffi                      1.12.3           py37h2e261b9_0  
chardet                   3.0.4                    pypi_0    pypi
cloudpickle               1.2.1                      py_0  
cudatoolkit               10.0.130                      0  
cycler                    0.10.0                     py_1    conda-forge
cytoolz                   0.10.0           py37h7b6447c_0  
dask-core                 2.2.0                      py_0  
dbus                      1.13.6               h746ee38_0  
decorator                 4.4.0                    py37_1  
expat                     2.2.6                he6710b0_0  
ffmpeg                    4.0                  hcdf2ecd_0  
fontconfig                2.13.0               h9420a91_0  
freeglut                  3.0.0                hf484d3e_5  
freetype                  2.9.1                h8a8886c_1  
geos                      3.7.2                he1b5a44_1    conda-forge
glib                      2.56.2               hd408876_0  
glob2                     0.7                      pypi_0    pypi
graphite2                 1.3.13               h23475e2_0  
gst-plugins-base          1.14.0               hbbd80ab_1  
gstreamer                 1.14.0               hb453b48_1  
harfbuzz                  1.8.8                hffaf4a1_0  
hdf5                      1.10.2               hba1933b_1  
icu                       58.2                 h9c2bf20_1  
idna                      2.8                      pypi_0    pypi
imageio                   2.5.0                    py37_0  
imgaug                    0.2.9                      py_1    conda-forge
intel-openmp              2019.4                      243  
jasper                    2.0.14               h07fcdf6_1  
joblib                    0.13.2                   py37_0    anaconda
jpeg                      9b                   h024ee3a_2  
kiwisolver                1.1.0            py37hc9558a2_0    conda-forge
libedit                   3.1.20181209         hc058e9b_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 9.1.0                hdf63c60_0  
libgfortran-ng            7.3.0                hdf63c60_0  
libglu                    9.0.0                hf484d3e_1  
libopencv                 3.4.2                hb342d67_1  
libopus                   1.3                  h7b6447c_0  
libpng                    1.6.37               hbc83047_0  
libprotobuf               3.9.1                h8b12597_0    conda-forge
libstdcxx-ng              9.1.0                hdf63c60_0  
libtiff                   4.0.10               h2733197_2  
libuuid                   1.0.3                h1bed415_2  
libvpx                    1.7.0                h439df22_0  
libxcb                    1.13                 h1bed415_1  
libxml2                   2.9.9                hea5a465_1  
markdown                  3.1.1                      py_0    conda-forge
matplotlib                3.1.0            py37h5429711_0  
mkl                       2019.4                      243  
mkl-service               2.0.2            py37h7b6447c_0  
mkl_fft                   1.0.12           py37ha843d7b_0  
mkl_random                1.0.2            py37hd81dba3_0  
munch                     2.3.2                    pypi_0    pypi
ncurses                   6.1                  he6710b0_1  
networkx                  2.3                        py_0  
ninja                     1.9.0            py37hfd86e86_0  
numpy                     1.16.4           py37h7e9f1db_0  
numpy-base                1.16.4           py37hde5b4d6_0  
olefile                   0.46                     py37_0  
opencv                    3.4.2            py37h6fd60c2_1  
openssl                   1.1.1c               h516909a_0    conda-forge
pandas                    0.25.0           py37he6710b0_0  
pcre                      8.43                 he6710b0_0  
pillow                    6.1.0            py37h34e0f95_0  
pip                       19.1.1                   py37_0  
pixman                    0.38.0               h7b6447c_0  
pretrainedmodels          0.7.4                    pypi_0    pypi
protobuf                  3.9.1            py37he1b5a44_0    conda-forge
py-opencv                 3.4.2            py37hb342d67_1  
pycparser                 2.19                     py37_0  
pydicom                   1.3.0                      py_0    conda-forge
pyparsing                 2.4.2                      py_0    conda-forge
pyqt                      5.9.2            py37hcca6a23_2    conda-forge
python                    3.7.3                h0371630_0  
python-dateutil           2.8.0                    py37_0  
pytorch                   1.1.0           py3.7_cuda10.0.130_cudnn7.5.1_0    pytorch
pytz                      2019.1                     py_0  
pywavelets                1.0.3            py37hdd07704_1  
pyzmq                     18.0.2                   pypi_0    pypi
qt                        5.9.7                h5867ecd_1  
readline                  7.0                  h7b6447c_5  
requests                  2.22.0                   pypi_0    pypi
retrying                  1.3.3                      py_2    conda-forge
scikit-image              0.15.0           py37he6710b0_0  
scikit-learn              0.21.2           py37hd81dba3_0    anaconda
scipy                     1.3.0            py37h7c811a0_0  
segmentation-models-pytorch 0.0.1                    pypi_0    pypi
setuptools                41.0.1                   py37_0  
shapely                   1.6.4           py37hec07ddf_1006    conda-forge
sip                       4.19.8          py37hf484d3e_1000    conda-forge
six                       1.12.0                   py37_0  
sqlite                    3.29.0               h7b6447c_0  
tensorboard               1.14.0                   py37_0    conda-forge
tensorboardx              1.8                        py_0    conda-forge
tk                        8.6.8                hbc83047_0  
toolz                     0.10.0                     py_0  
torchcontrib              0.0.2                    pypi_0    pypi
torchfile                 0.1.0                    pypi_0    pypi
torchnet                  0.0.4                    pypi_0    pypi
torchvision               0.3.0           py37_cu10.0.130_1    pytorch
tornado                   6.0.3            py37h516909a_0    conda-forge
tqdm                      4.32.1                     py_0  
urllib3                   1.25.3                   pypi_0    pypi
visdom                    0.1.8.8                  pypi_0    pypi
websocket-client          0.56.0                   pypi_0    pypi
werkzeug                  0.15.5                     py_0    conda-forge
wheel                     0.33.4                   py37_0  
xz                        5.2.4                h14c3975_4  
zlib                      1.2.11               h7b6447c_3  
zstd                      1.3.7                h0b5b093_0  
